test: consolidate implicit workspace entry regression
